refactor: move components in blog-page to `components/` folder Refs #6
