+++ src/components/blog/hero-card.tsx
